requirements.txt
# Core dependencies (required)
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
Pillow>=8.0.0
opencv-python>=4.5.0
PyYAML>=6.0

# For DiT models (required)
timm>=0.9.0

# For diffusion (required)
scipy>=1.7.0

# For VAE (required)
diffusers>=0.21.0

# For training (required)
lightning>=2.0.0
tensorboard>=2.8.0

# For monitoring and plotting (required)
matplotlib>=3.5.0

# For logging (optional)
wandb>=0.15.0